Feature market/nam lv/approve : fix bug put on sale

See merge request quannv_edsolabs/dfy-mobile!809

+++ lib/presentation/put_on_market/ui/pawn_tab.dart
@@ -182,7 +182,7 @@ class _PawnTabState extends State<PawnTab>
                 },
                 onchangeText: (value) {
                   widget.cubit.changeTokenPawn(
-                    value: value != '' ? int.parse(value) : 0,
+                    value: value != '' ? double.parse(value) : 0,
                   );
                   _putOnMarketModel.price = value;
                 },
